# docs/openapi.yml
openapi: 3.0.0
info:
  title: Rudex
  description: >
    Authentication service
  version: "0.1"
  x-displayName: Authentication service

servers:
  - url: https://app.42pasila.vn/rudex
    description: Production

paths:
  /login:
    post:
      operationId: loginUser
      summary: Login a user and return authentication tokens
      description: >
        Authenticates a user using email/username and password, returning
        a short-lived access token and a refresh token.
      tags:
        - User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestBody"
      responses:
        "200":
          description: Successful login response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseBody"
        "400":
          $ref: "#/components/responses/APIErrorResponse"
        "401":
          $ref: "#/components/responses/APIErrorResponse"
        "500":
          $ref: "#/components/responses/APIErrorResponse"

  /register:
    post:
      operationId: registerUser
      summary: Register a user
      description: >
        Create a new user with provided input.
      tags:
        - User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequestBody"
      responses:
        "201":
          description: New user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponseBody"
        "400":
          $ref: "#/components/responses/APIErrorResponse"
        "409":
          $ref: "#/components/responses/APIErrorResponse"
        "500":
          $ref: "#/components/responses/APIErrorResponse"
  /confirm:
    get:
      operationId: confirmEmail
      summary: Confirm a user's email address
      description: >
        Validates the confirmation token from the email link and marks
        the user's email as confirmed.
      tags:
        - Email confirmation
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
          description: The confirmation token sent via email
      responses:
        "200":
          description: Email confirmed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmEmailResponseBody"
        "400":
          $ref: "#/components/responses/APIErrorResponse"
        "409":
          $ref: "#/components/responses/APIErrorResponse"
        "500":
          $ref: "#/components/responses/APIErrorResponse"

  /users/usernames:
    get:
      operationId: getUserNames
      summary: Get all user names
      description: >
        Get user names with optional filtering by user IDs and pagination.
      tags:
        - Users
      parameters:
        - in: query
          name: rudexUserIds
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/rudexUserId"
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-indexed)
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
          description: Number of items per page
      responses:
        "200":
          $ref: "#/components/responses/GetUserNamesResponseBody"
        "400":
          $ref: "#/components/responses/APIErrorResponse"
        "404":
          $ref: "#/components/responses/APIErrorResponse"
        "500":
          $ref: "#/components/responses/APIErrorResponse"

components:
  parameters:
    rudexUserIdQuery:
      name: rudexUserId
      in: query
      description: Unique ID of a Rudex user (optional)
      required: false
      schema:
        $ref: "#/components/schemas/rudexUserId"

  schemas:
    rudexUserId:
      type: string
      description: Rudex user identifier (string). Use UUID if you require a UUID format.
      example: "123e4567-e89b-12d3-a456-426614174000"

    LoginRequestBody:
      type: object
      additionalProperties: false
      required:
        - username
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        googleUserId:
          type: string

    LoginResponseBody:
      type: object
      required:
        - accessToken
        - accessTokenExpiryDate
        - refreshToken
        - userId
      properties:
        accessToken:
          type: string
          description: JWT access token used for authenticated requests.
        accessTokenExpiryDate:
          type: string
          format: date-time
          description: Expiration date and time of the access token.
        refreshToken:
          type: string
          description: Token used to refresh access tokens.
        userId:
          type: string

    RegisterRequestBody:
      type: object
      additionalProperties: false
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string

    RegisterResponseBody:
      type: object
      additionalProperties: false
      required:
        - rudexUserId
      properties:
        rudexUserId:
          $ref: "#/components/schemas/rudexUserId"

    ConfirmEmailResponseBody:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string

    GetUserNamesResponseBody:
      type: object
      additionalProperties: false
      required:
        - users
        - total
        - page
        - limit
      properties:
        users:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - username
              - id
            properties:
              username:
                type: string
              id:
                $ref: "#/components/schemas/rudexUserId"
        total:
          type: integer
          description: Total number of matching users
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
  responses:
    GetUserNamesResponseBody:
      description: Successful response with user names
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GetUserNamesResponseBody"
    APIErrorResponse:
      description: API error response
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              type:
                type: string
              message:
                type: string
              stack:
                type: string
              info:
                type: object